+++ octasine/src/gui/operator.rs
@@ -4,6 +4,7 @@ use iced_baseview::{
     widget::Text, Alignment, Element, Length,
 };
 
+use crate::parameters::velocity_sensitivity::VelocitySensitivityValue;
 use crate::parameters::{
     Operator2ModulationTargetValue, Operator3ModulationTargetValue, Operator4ModulationTargetValue,
     OperatorFeedbackValue, OperatorFrequencyFineValue, OperatorFrequencyFreeValue,
@@ -30,6 +31,7 @@ pub enum ModTargetPicker {
 
 pub struct OperatorWidgets {
     index: usize,
+    pub shifted: bool,
     pub volume: OctaSineKnob<OperatorVolumeValue>,
     pub mute_button: BooleanButton,
     pub mix: OctaSineKnob<OperatorMixOutValue>,
@@ -41,6 +43,8 @@ pub struct OperatorWidgets {
     pub frequency_ratio: OctaSineKnob<OperatorFrequencyRatioValue>,
     pub frequency_free: OctaSineKnob<OperatorFrequencyFreeValue>,
     pub frequency_fine: OctaSineKnob<OperatorFrequencyFineValue>,
+    pub mod_out_velocity_sensitivity: OctaSineKnob<VelocitySensitivityValue>,
+    pub feedback_velocity_sensitivity: OctaSineKnob<VelocitySensitivityValue>,
     pub envelope: Envelope,
     pub wave_display: WaveDisplay,
 }
@@ -71,6 +75,7 @@ impl OperatorWidgets {
 
         Self {
             index: operator_index,
+            shifted: false,
             volume: knob::operator_volume(sync_handle, operator_index),
             mute_button: operator_mute_button(sync_handle, operator_index),
             mix: knob::operator_mix(sync_handle, operator_index),
@@ -84,6 +89,14 @@ impl OperatorWidgets {
             frequency_fine: knob::operator_frequency_fine(sync_handle, operator_index),
             envelope: Envelope::new(sync_handle, operator_index),
             wave_display: WaveDisplay::new(sync_handle, operator_index),
+            mod_out_velocity_sensitivity: knob::operator_mod_out_velocity_sensitivity(
+                sync_handle,
+                operator_index,
+            ),
+            feedback_velocity_sensitivity: knob::operator_feedback_velocity_sensitivity(
+                sync_handle,
+                operator_index,
+            ),
         }
     }
 
@@ -178,8 +191,20 @@ impl OperatorWidgets {
                 .push(container_l3(self.frequency_fine.view(theme))),
         );
 
-        let envelope = container_l2(self.envelope.view(theme))
-            .height(Length::Fixed(f32::from(LINE_HEIGHT * 8)));
+        let end = if self.shifted {
+            container_l2(
+                Row::new()
+                    .push(space_l3())
+                    .push(container_l3(self.mod_out_velocity_sensitivity.view(theme)))
+                    .push(space_l3())
+                    .push(container_l3(self.feedback_velocity_sensitivity.view(theme)))
+                    .push(space_l3().width(LINE_HEIGHT * 15)),
+            )
+        } else {
+            container_l2(self.envelope.view(theme))
+                .height(Length::Fixed(f32::from(LINE_HEIGHT * 8)))
+                .into()
+        };
 
         container_l1(
             Row::new()
@@ -190,7 +215,7 @@ impl OperatorWidgets {
                 .push(space_l2())
                 .push(frequency_group)
                 .push(space_l2())
-                .push(envelope),
+                .push(end),
         )
         .into()
     }